requirements.txt
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.9.0
biopython==1.76
certifi==2020.4.5
chardet==3.0.4
ConfigArgParse==1.2
coverage==5.0.4
cycler==0.10.0
datrie==0.8.2
decorator==4.4.2
docutils==0.16
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
importlib-metadata==1.6.0
ipython-genutils==0.2.0
jsonschema==3.2.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.0
matplotlib==3.2.1
more-itertools==8.2.0
nbformat==5.0.5
nltk==3.2.4
numpy==1.18.2
packaging==20.3
pluggy==0.13.1
preprocessing==0.1.13
psutil==5.7.0
py==1.8.1
pyparsing==2.4.6
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.1
PyYAML==5.3.1
ratelimiter==1.2.0.post0
requests==2.23.0
six==1.14.0
smmap==3.0.1
snakemake==5.13.0
soupsieve==2.0
sphinx-rtd-theme==0.2.4
toposort==1.5
traitlets==4.3.3
uniprot==1.3
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.12.1
zipp==3.1.0
